Make feed assertions available outside core gem
