+++ publify_core/Manifest.txt
@@ -345,6 +345,7 @@ lib/publify_core.rb
 lib/publify_core/engine.rb
 lib/publify_core/lang.rb
 lib/publify_core/testing_support/factories.rb
+lib/publify_core/testing_support/feed_assertions.rb
 lib/publify_core/version.rb
 lib/publify_guid.rb
 lib/publify_plugins.rb
